requirements.txt
labelbox
python-dotenv